Rename `opdef` to `op_def` for consistency with `type_def`.
